+++ Runtime/Conditions/ComparisonOperator.cs
@@ -1,0 +1,10 @@
+// Copyright 2026 Spellbound Studio Inc.
+
+namespace Spellbound.Modifiers {
+    public enum ComparisonOperator {
+        LessThan,
+        AtMost,
+        AtLeast,
+        GreaterThan
+    }
+}