+++ Tests/Modifiers/ModifierGrantSetTests.cs
@@ -1,6 +1,7 @@
 // Copyright 2026 Spellbound Studio Inc.
 
 using System;
+using System.Collections.Generic;
 using NUnit.Framework;
 using UnityEngine;